Add files via upload
Sorry if the code is too random first I created another file that included a rectangle and ellipse moving randomly with random colors. Last-minute I changed my mind and I tried creating a loop that was supposed to contain one rectangle moving in the vertical direction and another in horizontal but I messed it up. I will try to re-upload a new file by the end of the week once I learn how to create loops.