Add files via upload
For this exercise, I was aiming to create a background similar to the lofi music youtube videos (for ex.https://www.youtube.com/watch?v=q55qNEKQLG0&ab_channel=LofiMusic) where the background is still but there is a rain effect on it. However, I wanted to do that with a panoramic effect on the picture that I learned during the lecture. Unfortunately, I was not able to find an anime panoramic picture that represents city streets and my choices were very limited for the background. I was able to create the rain effect (reference in the code) and perform the panoramic effect on the image. On top of that, I created a character that I was hoping I could make stay still in one place or add a moving effect to it but I still haven't learned that effect. I hope you like the results.