Project_1
This game is called the Cannon game. The aim of this game is to fire as many blue balls as possible vertically before the balls hit the ground. To play the game, you can move the cannon horizontally through the right and left key on the keyboard and you can fire the cannonball by pressing the space key which releases the cannonball each time is pressed.

Please note this is the updated version and the first version was updated 14 days but unfortunately I deleted it when attempt to update the files.